Timestamp file names in downloads folder
Add a timestamp to the file names copied to the downloads
folder for scalability. Each download adds a timestamp to the
file name so that there's no concurrent access to the file being
downloaded. I also made the file search async by converting
to a list due to an EF exception to prevent infinite streaming.
The results here need to be paged.